draft

+++ matlab/util/+tools/+array/mex_bitand_array64.c
@@ -1,4 +1,4 @@
-typedef unsigned short __Int;
+typedef unsigned long long __Int;
 #define MEX_INT  mxINT16_CLASS
 #define MEX_UINT mxUINT16_CLASS
